% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType_genes_jitter.R
\name{CellType_genes_jitter}
\alias{CellType_genes_jitter}
\title{Plot gene expression for all CellTypes as jitter plots}
\usage{
CellType_genes_jitter(gene_to_plot, cds, cell_size = 0.75, nrow = NULL,
  ncol = 2, color = "genotype", group = "genotype", plot_trend = F,
  relative_expr = F, color_scale = c("Black", "Red"))
}
\arguments{
\item{gene_to_plot}{official gene symbol of a gene to plot e.g. "Actb"}

\item{cds}{a CellDataSet object e.g. used in the monocle package}

\item{cell_size}{the size (in points) of each cell used in the plot.}

\item{nrow}{the number of rows used when laying out the panels for each gene's expression.}

\item{ncol}{the number of columns used when laying out the panels for each gene's expression.}

\item{color}{the cell attribute (e.g. the column of pData(cds)) to be used to color each cell.}

\item{group}{the cell attribute (e.g. the column of pData(cds)) to group cells by on the horizontal axis.}

\item{plot_trend}{whether to plot a trendline tracking the average expression across the horizontal axis.}

\item{relative_expr}{Whether to transform expression into relative values.}

\item{color_scale}{a vector of colors to color barplots and error bars with.}
}
\value{
A ggplot2 object
}
\description{
CellType_genes_jitter is a wrapper function around monocle::plot_genes_jitter() that plots gene expression
as jitter plots facet_wrapped by CellType.
}
\examples{
gene_barplot("Actb", dat)

}
\keyword{gene_barplot()}
