% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monocle_cds.R
\name{create_monocle_cds}
\alias{create_monocle_cds}
\title{Create a monocle CellDataSet object from a directory of cellranger-aligned 10X Genomics scRNAseq data}
\usage{
create_monocle_cds(cellranger_outs_path, cellranger_filter = F,
  UMI_cutoff = NULL, testDrops = F, lower_UMI_threshold = 100,
  upper_UMI_threshold = 500, FDR = 0.01, cell_barcodes = NULL)
}
\arguments{
\item{cellranger_outs_path}{the path to the "outs" directory in the cellranger library folder e.g.
"/mycellranger_library/outs"}

\item{cellranger_filter}{whether to use the filtered matrices from the cellranger pipeline as the 
expression matrix (i.e. as cells)}

\item{UMI_cutoff}{numeric; a single cutoff of total UMI counts per cell to use for determining cells (i.e. 
all barcodes with total UMI counts above this cutoff will be included in the expression matrix); defaults 
to NULL}

\item{testDrops}{logical; whether to use the testDrops function to determine which barcodes are cells (and
therefore included in the expression matrix)}

\item{lower_UMI_threshold}{numeric; parameter for the testDrops function. All barcodes with total UMI counts
below this value are considered empty droplets (i.e. not cells)}

\item{upper_UMI_threshold}{numeric; parameter for the testDrops function. All barcodes with total UMI counts
above this value are considered cells}

\item{FDR}{numeric; parameter for the testDrops function. FDR value to use in the testDrops function when
statistically determining cells vs non-cells}

\item{cell_barcodes}{a vector of barcodes to use for creating the expression matrix (all barcodes specified
by cell_barcodes are assumed to be cells)}
}
\value{
a monocle CellDataSet object
}
\description{
create_monocle_cds() creates a monocle CellDataSet object from a directory of 10X Genomics scRNAseq data
created using the cellranger pipeline. Includes options to determine which barcodes are cells vs non-cells
in the original raw matrix.
}
\examples{
dat <- create_monocle_cds(cellranger_outs_path, cellranger_filter= F, UMI_cutoff=NULL, testDrops=T,
lower_UMI_threshold=100, upper_UMI_threshold=500, FDR=0.01, cell_barcodes=NULL)
}
\keyword{create_monocle_cds}
