% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultiSoupChannel.R
\name{createMultiSoupChannel}
\alias{createMultiSoupChannel}
\title{Create multiple Soup Channels from an cellranger-aggregated matrix}
\usage{
createMultiSoupChannel(cds, AllBlankDrops, channelNames = NULL,
  soupRange = c(0, 10))
}
\arguments{
\item{cds}{a CellDataSet object (e.g. used in the monocle package)}

\item{AllBlankDrops}{an expression matrix of empty droplets from all samples in the cellranger aggregate.
Matrix must contain barcodes from all the different samples used to form the cellranger aggregate; this
can be derived from the original aggregated raw matrix using the testDrops function. Expression values
from the AllBlankDrops matrix are used to estimate the soup.}

\item{channelNames}{vector of strings; channel names to use for all the channels in the sample; number
of channel names must equal number of samples used in the cellranger aggregate}

\item{soupRange}{Droplets with total UMI count in this range (excluding endpoints) are used to estimate soup.}
}
\value{
a SoupChannel object with multiple channels (corresponding to number of samples)
}
\description{
createMultiSoupChannel() is a wrapper function around the createSoupChannel() function that creates
a channel for each sample in a matrix aggregated from multiple samples with the cellranger pipeline.
The function requires a AllBlankDrops object, which is a matrix consisting of the empty droplets from
all samples derived from the original cellranger-aggregated raw matrix using the testDrops() function.
Based on the excellent SoupX package.
}
\examples{
scl <- createMultiSoupChannel(cds=dat, AllBlankDrops=raw_mtx[,all_blank_barcodes], channelNames=NULL
soupRange=c(0,10))
}
\keyword{createMultiSoupChannel}
