% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSeuratObjectList.R
\name{CreateSeuratObjectList}
\alias{CreateSeuratObjectList}
\title{Create a list of Seurat Objects}
\usage{
CreateSeuratObjectList(exprs_matrix, phenoData)
}
\arguments{
\item{exprs_matrix}{an expression matrix; row names must be gene names and column names must be cell barcodes}

\item{phenoData}{dataframe where rows are cells and rownames are cell barcodes, columns are fields defining 
cellular characteristics. Row names of phenoData must match column names of exprs_matrix.}
}
\value{
A list of Seurat Objects
}
\description{
CreateSeuratObjectList() creates a list of Seurat objects for use in the RunMultiCCA function in the R package
Seurat. The function splits a single aggregated barcoded expression matrix into separate matries based on the 
suffix at the end of each cellular barcode e.g. ("ACTAGGAGACAGGTGC-1").
}
\examples{
list <- CreateSeuratObjectList(exprs_matrix = exprs(dat), phenoData = pData(dat))
}
\keyword{CreateSeuratObjectList}
