% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSoupChannel.R
\name{createSoupChannel}
\alias{createSoupChannel}
\title{Create a single Soup Channel}
\usage{
createSoupChannel(blankDrops, cells, channelName, soupRange = c(0, 10),
  keepDroplets = FALSE, ...)
}
\arguments{
\item{blankDrops}{an expression matrix of empty droplets derived from the original raw matrix using 
the testDrops() function. Expression values from the blankDrops matrix are used to estimate the soup using
the emptySoup() function.}

\item{cells}{an expression matrix of cells (e.g. determined by the testDrops() function)}

\item{channelName}{string; the name of the channel}

\item{soupRange}{Droplets with total UMI count in this range (excluding endpoints) are used to estimate soup.}

\item{keepDroplets}{logical; whether to keep the blankDrops matrix after estimating the soup; defaults to
FALSE}
}
\value{
a SoupChannel object
}
\description{
createSoupChannel is a modified version of the SoupChannel() function in the excellent SoupX package. Instead 
of using a table of droplets (tod) that include all droplets to estimate the "soup", createSoupChannel feeds 
an expression matrix consisting of only empty droplets, determined by the testDrops() function, to estimate
the soup in the sample.
}
\examples{
scl <- createSoupChannel(blankDrops=raw_mtx[,blank_barcodes], cells=raw_mtx[,cell_barcodes], channelNames=NULL
soupRange=c(0,10))
}
\keyword{createSoupChannel}
