% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptySoup.R
\name{emptySoup}
\alias{emptySoup}
\title{Estimate expression profile of soup from empty droplets}
\usage{
emptySoup(sc, soupRange = c(0, 10), keepDroplets = FALSE)
}
\arguments{
\item{sc}{a SoupChannel object created using the createSoupChannel() function implemented in cellwrangler.}

\item{soupRange}{Droplets with total UMI count in this range (excluding endpoints) are used to estimate soup.}

\item{keepDroplets}{logical; whether to keep the blankDrops matrix after estimating the soup; defaults to
FALSE}
}
\value{
a SoupChannel object
}
\description{
emptySoup() is a modified version of the estimateSoup() function in the excellent SoupX package. Instead
of using all droplets within the specified soupRange to estimate the expression profile of the
soup as per the estimateSoup() function, emptySoup() uses only empty droplets e.g. determined using 
the testDrops() function to estimate the expression profile of the soup.
}
\examples{
scl <- emptySoup(scl, soupRange = c(0, 10), keepDroplets = FALSE)
}
\keyword{emptySoup}
