% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_threshold.R
\name{exprs_threshold}
\alias{exprs_threshold}
\title{Determine if cells are above an expression threshold of a gene}
\usage{
exprs_threshold(cds, gene_name, threshold_value)
}
\arguments{
\item{cds}{a monocle CellDataSet object.}

\item{gene_name}{official gene name e.g. "Actb".}

\item{threshold_value}{numeric; threshold of expression of gene.}
}
\value{
A boolean vector
}
\description{
dim_coord_barcode() returns the rownames of cells based on the coordiates of a dimensionality reduction
projection (e.g. t-SNE, UMAP).
}
\examples{
exprs_threshold(dat, "Actb", 1)
}
\keyword{exprs_threshold}
