% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_CCA_genes.R
\name{find_CCA_genes}
\alias{find_CCA_genes}
\title{Find high variance genes for CCA}
\usage{
find_CCA_genes(SeuratObjectList, num_hvg_genes = 1000,
  num_samples_variable = 2)
}
\arguments{
\item{SeuratObjectList}{a list of Seurat objects}

\item{num_hvg_genes}{number of high-variance genes per sample to examine starting from the most variable}

\item{num_samples_variable}{minimum numher of samples the high-variance genes must be highly variable in}
}
\value{
A list of Seurat Objects
}
\description{
find_CCA_genes() defines genes to use for CCA by finding high-variance genes in a list of Seurat objects
that are highly variable in a minimum number of samples/objects.
}
\examples{
genes.use <- find_CCA_genes(SeuratObList,  num_hvg_genes=1000, num_samples_variable=2)
}
\keyword{CreateSeuratObjectList}
