% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sample_dispersed_genes.R
\name{find_sample_dispersed_genes}
\alias{find_sample_dispersed_genes}
\title{Find dispersed genes in each sample in CellDataSet}
\usage{
find_sample_dispersed_genes(cds, mean_exprs = 0, disp_empirical = 1)
}
\arguments{
\item{cds}{a monocle CellDataSet object}

\item{mean_exprs}{mean expression threshold for defining dispersed genes}

\item{disp_empirical}{dispersion empirical threshold for defining dispersed genes}
}
\value{
A list of vectors with each vector containing dispersed genes for each sample
}
\description{
find_sample_dispersed_genes() splits an aggregated cds into separate samples based on the suffix at the end 
of each barcode and estimates SizeFactors and Dispersions for each sample using the package monocle. The
function then proceeds to extract genes above pre-specified mean expression and dispersion empirical
thresholds
}
\examples{
find_sample_dispersed_genes(dat, mean_exprs=0, disp_empirical=1)
}
\keyword{find_sample_dispersed_genes}
