% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sample_expressed_genes.R
\name{find_sample_expressed_genes}
\alias{find_sample_expressed_genes}
\title{Find expressed genes in each sample in CellDataSet}
\usage{
find_sample_expressed_genes(cds, UMI_threshold = 1, cell_threshold = 1)
}
\arguments{
\item{cds}{a monocle CellDataSet object}

\item{UMI_threshold}{UMI threshold for defining expressed genes. Genes with total UMIs equal to or above
this threshold will be selected.}

\item{cell_threshold}{cell threshold for defining expressed genes. Genes expressed in a total number of cells 
at or above this threshold will be selected.}
}
\value{
A list of vectors with each vector containing dispersed genes for each sample
}
\description{
find_sample_expressed_genes() splits an aggregated cds into separate samples based on the suffix at the 
end of each barcode. The function then proceeds to extract genes above the pre-specified UMI and cell 
expression threshold(s)
}
\examples{
find_sample_expressed_genes(dat,UMI_threshold = 1, cell_threshold = 10)
}
\keyword{find_sample_expressed_genes}
