% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGeneID.R
\name{findGeneID}
\alias{findGeneID}
\title{Find Ensembl gene IDs given gene names}
\usage{
findGeneID(gene_names, cds, unique = T)
}
\arguments{
\item{gene_names}{a vector of official gene names e.g. "Acta2"}

\item{cds}{a CellDataSet object e.g. used in the monocle package}
}
\description{
Using gene names as input, findGeneID() returns the corresponding Ensembl gene IDs in the fData
of the specified CellDataSet object.
}
\examples{
findGeneID(c("Acta2","Cdh5","Lum"), myCDS)
}
\keyword{findGeneID}
