% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRows.R
\name{findRows}
\alias{findRows}
\title{Find rows by row names}
\usage{
findRows(row_names, matrix)
}
\arguments{
\item{row_names}{a vector of rowname(s) of row(s) to return}

\item{matrix}{matrix or dataframe}
}
\description{
findRows() return specified rows in a matrix or dataframe by specifying the corresponding rownames.
}
\examples{
findRows("myrowname", mydataframe)
}
\keyword{findRows}
