% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_barplot.R
\name{gene_barplot}
\alias{gene_barplot}
\title{Plot mean gene expression as bar plot}
\usage{
gene_barplot(genes, cds, group = "genotype", color = "genotype",
  facet_wrap = NULL, facet_genes = "cols", plot_trend = F,
  color_trend = "orange")
}
\arguments{
\item{genes}{a vector of gene name(s) to plot e.g. c("Actb", "Aldoa")}

\item{cds}{a CellDataSet object e.g. used in the monocle package}

\item{group}{column name in pData(cds) to group the bar plots by on the horizontal axis.}

\item{color}{column name in pData(cds) to color the bar plots with.}

\item{facet_wrap}{a column in pData(cds) to facet_wrap the plots with.}

\item{facet_genes}{only used if facet_wrap is NULL and there are multiple genes being plotted; if set to "cols", will
facet genes into columns; otherwise will facet genes into rows}

\item{plot_trend}{logical;whether to plot a trend line across groups in plot(s).}

\item{color_trend}{color to use for trend line.}
}
\value{
A ggplot2 object
}
\description{
gene_barplot() plots the mean expression of a given gene as a bar plot with standard error bars.
}
\examples{
gene_barplot(c("Actb", "Aldoa"), dat)
}
\keyword{gene_barplot()}
