% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_jitterplot.R
\name{gene_jitterplot}
\alias{gene_jitterplot}
\title{Plot expression of cells as jitter plot}
\usage{
gene_jitterplot(genes, cds, group = "genotype", color = NULL,
  cell_size = 0.75, exprs_threshold = 0, nrow = NULL, ncol = 1,
  panel_order = NULL, gene_groups = NULL, plot_trend = FALSE,
  color_trend = "orange", label_by_short_name = TRUE,
  log_expr = TRUE, relative_expr = FALSE, scales = "free_y")
}
\arguments{
\item{genes}{a vector of gene name(s) to plot e.g. c("Actb", "Aldoa")}

\item{cds}{a CellDataSet object e.g. used in the monocle package}

\item{group}{column name in pData(cds) to group cells by on the horizontal axis.}

\item{color}{column name in pData(cds) to color cells with.}

\item{cell_size}{the size (in points) of each cell used in the plot}

\item{exprs_threshold}{the (untransformed) expression threshold to use in plotting genes; expression 
values less than or equal to this threshold will be converted to 0}

\item{nrow}{the number of rows used when laying out the panels for each gene's expression}

\item{ncol}{the number of columns used when laying out the panels for each gene's expression}

\item{panel_order}{the order in which genes should be layed out (left-to-right, top-to-bottom)}

\item{gene_groups}{dataframe specifying groups for genes to be plotted; first column should be gene id or 
gene short name (if label_by_short_name = T), and second column should be group id for gene; can be used
for facet_wrap or color by using the variable "gene_group"}

\item{plot_trend}{logical;whether to plot a trend line across groups in plot(s).}

\item{color_trend}{color to use for trend line.}

\item{label_by_short_name}{logical;label figure panels by gene_short_name (TRUE) or feature id (FALSE)}

\item{log_expr}{logical; whether to log-transform expression into log10(expression + 1)}

\item{relative_expr}{logical;Whether to transform expression into counts per 10,000}

\item{scales}{input for facet_wrap scale; options include "free_y" or "fixed"}
}
\value{
A ggplot2 object
}
\description{
gene_jitterplot() plots the expression of a given gene as a jitter plot for a CellDataSet object.
}
\examples{
gene_jitterplot(c("Actb", "Aldoa"), cds)
}
\keyword{gene_jitterplot}
