% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expressed_genes.R
\name{get_expressed_genes}
\alias{get_expressed_genes}
\title{Find genes expressed in a CellDataSet object}
\usage{
get_expressed_genes(cds, cell_threshold = 1, min_prop = NULL,
  min_condition = NULL)
}
\arguments{
\item{cds}{a CellDataSet object used in the monocle package}

\item{cell_threshold}{numerical threshold for number of cells that must express a particular gene e.g. 1; 
if min_prop is specified, this number will be overridden.}

\item{min_prop}{minimum proportion of cells that must express a particular gene e.g. 0.05 for the gene
to be considered expressed}

\item{min_condition}{a column in pData(cds) that specifies a condition e.g. genotype to which the min_prop 
argument will be applied; must be specified together with the min_prop argument. Takes the minimum proportion
of the condition with fewer cells as the threshold for finding expressed genes.}
}
\value{
CellDataSet object
}
\description{
get_expressed_genes() finds genes expressed in a CellDataSet object according to defined
parameters.
}
\examples{
expressed_genes <- get_expressed_genes(dat, min_prop=0.05, min_condition="genotype")
}
\keyword{get_expressed_genes}
