% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg_database.R
\name{get_kegg_database}
\alias{get_kegg_database}
\title{Get kegg database for an organism}
\usage{
get_kegg_database(database, organism)
}
\arguments{
\item{database}{a KEGG database; one of options available via KEGGREST listDatabases()}

\item{organism}{KEGG organism code e.g. "mmu" for "Mus musculus"}
}
\value{
a dataframe
}
\description{
get_kegg_database() is a wrapper function around the keggList() function that retrieves 
a kegg database for an organism and returns a dataframe
}
\examples{
kegg_db <- get_kegg_database("pathway","mmu")
}
\keyword{get_kegg_database}
