% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg_pathway_genes.R
\name{get_kegg_pathway_genes}
\alias{get_kegg_pathway_genes}
\title{Get genes for a kegg pathway}
\usage{
get_kegg_pathway_genes(pathway_name, kegg_pathway_df)
}
\arguments{
\item{pathway_name}{a KEGG pathway name}

\item{kegg_pathway_df}{a kegg pathway database that is produced using the cellwrangler get_kegg_database()
function using "pathway" as an option.}
}
\value{
a list
}
\description{
get_kegg_pathway_genes() is a wrapper function around keggGet() that retrieves gene names in a KEGG pathway.
}
\examples{
pathway_genes <- get_kegg_pathway_genes("Pentose phosphate pathway", mouse_kegg_pathways)
}
\keyword{get_kegg_database}
