% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_condition_counts.R
\name{get_mean_condition_counts}
\alias{get_mean_condition_counts}
\title{Get mean read counts for cells belonging to a condition}
\usage{
get_mean_condition_counts(cds, condition, read_type = "UMIs")
}
\arguments{
\item{cds}{a CellDataSet object used in the monocle package}

\item{condition}{column name in pData(cds) that specifies the condition for classifying cells e.g. genotype}

\item{read_type}{string specifying type of reads in expression matrix; defaults to "UMIs"}
}
\value{
CellDataSet object
}
\description{
get_mean_condition_counts finds the mean counts for cells belonging to a condition e.g. genotype 
in a CellDataSet object and returns the values in the pData of CellDataSet object
}
\examples{
dat <- get_mean_condition_counts(dat, condition="genotype")
}
\keyword{get_mean_condition_counts}
