% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patternMarkerNames.R
\name{get_patternMarkerNames}
\alias{get_patternMarkerNames}
\title{Get pattern markers as gene names for a set of CoGAPS results}
\usage{
get_patternMarkerNames(CoGAPS_res, cds, threshold = "all", lp = NA)
}
\arguments{
\item{CoGAPS_res}{a CoGAPS result object.}

\item{cds}{a CellDataSet object e.g. used in the monocle package}
}
\value{
a CoGAPS patternMarkers object
}
\description{
get_patternMarkerNames is a wrapper function around the CoGAPS::patternMarkers function for retrieving pattern 
markers as gene names for a set of CoGAPS results derived from a monocle CellDataSet object.
}
\examples{
patternMarkers <- get_patternMarkerNames(myCoGAPSres, cds=dat)
}
\keyword{get_patternMarkerNames}
