% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patternMarkers.R
\name{get_patternMarkers}
\alias{get_patternMarkers}
\title{Get pattern markers for a set of CoGAPS results}
\usage{
get_patternMarkers(CoGAPS_res, nFactor_range, relabel_genes = F,
  cds = NULL, scaledPmatrix = T, full = T)
}
\arguments{
\item{nFactor_range}{range of nFactor used in the set of CoGAPS results.}

\item{relabel_genes}{logical; if genes should be relabeled from Ensembl IDs to gene symbols.}

\item{cds}{CellDataSet object for relabeling genes.}

\item{CoGAPS_res_set}{a CoGAPS result set. If set to NULL, a Pmeans matrix may be supplied directly as input
for the Pattern_set parameter.}
}
\value{
a list of CoGAPS objects
}
\description{
get_patternMarkers is a wrapper function around the CoGAPS::patternMarkers function for retrieving pattern 
markers for a set of CoGAPS results.
}
\examples{
myCoGAPSres <- load_CoGAPS_res(res_dir = "myCoGAPSres_set", res_name = "myCoGAPS_result", 
nFactor_range = c(25,60,5))
}
\keyword{load_CoGAPS_res}
