% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_CoGAPS_res.R
\name{load_CoGAPS_res}
\alias{load_CoGAPS_res}
\title{Load a set of CoGAPS results}
\usage{
load_CoGAPS_res(res_dir, res_name, nFactor_range, name_output = T)
}
\arguments{
\item{res_dir}{name of directory containing CoGAPS results. Each CoGAPS result should be contained in
a folder with the name "run_x" in numerical order based on the length of nFactor_range.}

\item{res_name}{name of .RData result files.}

\item{nFactor_range}{range of nFactor used in the set of CoGAPS results.}

\item{name_output}{logical; whether to name output according to nFactor_range. Each CoGAPS_res will be 
named "nPXX", "nPXY" etc.}
}
\value{
a list of CoGAPS objects
}
\description{
load_CoGAPS_res() loads a set of CoGAPS results.
}
\examples{
myCoGAPSres <- load_CoGAPS_res(res_dir = "myCoGAPSres_set", res_name = "myCoGAPS_result", 
nFactor_range = c(25,60,5))
}
\keyword{load_CoGAPS_res}
