% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_normalize_cds.R
\name{log_normalize_cds}
\alias{log_normalize_cds}
\title{Normalize and log monocle cds}
\usage{
log_normalize_cds(cds, filter_zeros = TRUE, scale_to = "median",
  logbase = 10)
}
\arguments{
\item{cds}{a monocle CellDataSet object}

\item{filter_zeros}{logical - if genes with zero expression should be removed; defaults to TRUE}

\item{scale_to}{what value total read counts for each cell should be scaled to; by default, each cell's total
count is scaled to the median of all cells' totals; a numeric value can also be specified}

\item{logbase}{base for logarithm; defaults to 10; if NULL, no logarithm will be computed and the original
expression matrix will be returned}
}
\value{
a monocle CellDataSet object
}
\description{
log_normalize_cds() globally scales and log-transforms the expression matrix of a monocle CellDataSet object. 
The total count for each cell (represented by a column) is scaled to have the same total as all other cells 
(columns). A monocle cds with the log-normalized expression matrix is returned.
}
\examples{
dat_log <- log_normalize_cds(dat, scale_to = 10000, logbase=10)
}
\keyword{scale_log_cds}
