% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rnk.R
\name{make_rnk}
\alias{make_rnk}
\title{Make rank lists}
\usage{
make_rnk(celltype, rank_df, gene_name, rank_stat, decreasing = T,
  write_rnk = F, output_dir = NULL, output_name = NULL)
}
\arguments{
\item{celltype}{CellType to create rank lists for.}

\item{rank_df}{a dataframe with the pre-requisite rank statistics}

\item{gene_name}{column in rank_df containing gene names.}

\item{rank_stat}{column in rank_df containing statistic to rank genes by.}

\item{decreasing}{logical; whether to rank genes by decreasing value of rank_stat.}

\item{write_rnk}{logical; whether to write .rnk files for GSEA}

\item{output_dir}{name of directory to write .rnk files to; if directory does not exist, it will be created}

\item{output_name}{prefix for .rnk file names; .rnk files will be named "output_name_celltype.rnk"}

\item{group_column}{column name of column in annotation object with values corresponding to those 
in group_vector to use for computing logical columns. Defaults to NULL}
}
\value{
a list
}
\description{
make_rnk() makes rank lists for pre-ranked geneset enrichment analysis.
}
\examples{
make_rnk(Rods, rank_df, "Gene_symbol", "Rank stat", decreasing =T, write_rnk=F)
}
\keyword{make_rnk}
