% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scale_cds.R
\name{mean_scale_cds}
\alias{mean_scale_cds}
\title{Mean scale expression of cell groups across multiple samples to have the same mean in CellDataSet object}
\usage{
mean_scale_cds(cds, sample_col, group)
}
\arguments{
\item{cds}{monocle CellDataSet object}

\item{sample_col}{name of column in pData(cds) that annotates samples that cells are from}

\item{group}{name of column in pData(cds) that annotates groups that cells are in.}
}
\value{
A CellDataSet object
}
\description{
mean_scale_cds() scales the expression of cell groups e.g. CellType across multiple samples in
a monocle CellDataSet object such that cells from the same cell group in every sample have the same mean
expression. Assuming cells in same cell group have similar variances aross all samples, this will allow cells
from the same cell group in every sample to have similar distributions.
}
\examples{
dat <- mean_scale_cds(dat0, sample_col="sample", group="CellType")
}
\keyword{mean_scale_cds}
