% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_by_rownames.R
\name{merge_by_rownames}
\alias{merge_by_rownames}
\title{Merge dataframes or matrices by rownames}
\usage{
merge_by_rownames(x, y, all.x = TRUE, all.y = TRUE, sort = FALSE)
}
\arguments{
\item{x, y}{dataframes or matrices}

\item{all.x}{logical;  if TRUE, then extra rows will be added to the output, one for each row in x that has no 
matching row in y. These rows will have NAs in those columns that are usually filled with values from y. 
Defaults to TRUE}

\item{all.y}{logical; analogous to all.x}

\item{sort}{logical; if TRUE, result will be sorted by columns. Defaults to FALSE}
}
\description{
This is a wrapper function based on the base R merge function for merging 
two dataframes or matrices by their rownames and returing a dataframe or matrix with the appropriate rownames.
Unlike merge, this function does not create an additional column for rownames.
}
\examples{
merge_by_rownames(data.frame1, data.frame2)
}
\keyword{merge.by.rownames}
