% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df_list.R
\name{merge_df_list}
\alias{merge_df_list}
\title{Merge a list of dataframes or matrices}
\usage{
merge_df_list(df_list, all.x = TRUE, all.y = TRUE, sort = FALSE)
}
\arguments{
\item{df_list}{a list of dataframes or matrices}

\item{all.x}{logical;  if TRUE, then extra rows will be added to the output, one for each row in x that has no 
matching row in y. These rows will have NAs in those columns that are usually filled with values from y. 
Defaults to TRUE}

\item{all.y}{logical; analogous to all.x}

\item{sort}{logical; if TRUE, result will be sorted by columns. Defaults to FALSE}
}
\value{
A merged dataframe or matrix with the appropriate rownames
}
\description{
This is a wrapper function based on the merge_by_rownames function in cellwrangler for merging multiple
dataframes or matries by their rownames
}
\examples{
dataframes_list <- list(dataframe1,dataframe2,dataframe3,dataframe4)
merge_df_list(dataframes_list)
}
\keyword{merge_df_list}
