% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle_fitModel.R
\name{monocle_fitModel}
\alias{monocle_fitModel}
\title{Fit model function in monocle}
\usage{
monocle_fitModel(cds, modelFormulaStr = "~sm.ns(Pseudotime, df=3)",
  relative_expr = TRUE, cores = 1)
}
\arguments{
\item{cds}{a monocle CellDataSet object}

\item{modelFormulaStr}{model formula string specifying the model to fit for the genes}

\item{relative_expr}{Whether to fit a model to relative or absolute expression. Only meaningful for count-based 
expression data. If TRUE, counts are normalized by Size_Factor prior to fitting.}

\item{cores}{the number of processor cores to be used during fitting.}
}
\value{
a vglm object
}
\description{
monocle_fitModel() is the fitModel() function found in monocle 2. This function fits a vector 
generalized additive model (VGAM) from the VGAM package for each gene in a CellDataSet. By default, expression 
levels are modeled as smooth functions of the Pseudotime value of each cell. That is, expression is a function 
of progress through the biological process. More complicated formulae can be provided to account for additional 
covariates (e.g. day collected, genotype of cells, media conditions, etc).
}
\examples{
monocle_fitModel(dat, modelFormulaStr="~num_genes_expressed+genotype")
}
\keyword{monocle_fitModel()}
