% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_marker_matrix.R
\name{order_marker_matrix}
\alias{order_marker_matrix}
\title{Order expression matrix of marker genes by cell group}
\usage{
order_marker_matrix(cds, marker_genes, group_vector, scale = T,
  cluster_rows = T, cluster_cols = T,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  clustering_distance_cols = "euclidean",
  clustering_method_cols = "complete")
}
\arguments{
\item{cds}{CellDataSet object}

\item{marker_genes}{dataframe containing 2 columns: (1) gene symbol (2) cell group e.g. CellType}

\item{group_vector}{a vector classifying cells e.g. genotype or condition}

\item{scale}{"rows" or "cols" or "none"; indicates whether values should be centered and scaled in the 
row direction or the column direction or none.}

\item{cluster_rows}{logical; if rows should be clustered}

\item{cluster_cols}{logical; if columns should be clustered}

\item{clustering_distance_rows}{distance measure used in clustering rows. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method_rows}{clustering method used for rows. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{clustering_distance_cols}{distance measure used in clustering columns. Possible values are the same as
clustering_distance_rows.}

\item{clustering_method_cols}{clustering method used for columns. Accepts the same values as 
clustering_method_rows.}
}
\value{
A matrix
}
\description{
order_marker_matrix() returns an expression matrix of marker genes ordered by cell group
}
\examples{
order_marker_matrix(dat, marker_genes = myMarkers, group_vector = "CellType")
}
\keyword{order_marker_matrix()}
