% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bcv.R
\name{plot_bcv}
\alias{plot_bcv}
\title{Plot biological coefficent of variation and annotate genes}
\usage{
plot_bcv(annotated_fData, genes_to_color, color_scale = c("grey50",
  "red"), genes_to_annotate, title = NULL)
}
\arguments{
\item{annotated_fData}{annotated fData object produced by the cellwrangler annotate_bcv() function.}

\item{genes_to_color}{vector of Ensembl gene ids to color}

\item{color_scale}{color scale used for coloring genes specified in genes_to_color.}

\item{genes_to_annotate}{vector of Ensembl gene ids to annotate}
}
\description{
plot_bcv() plots the mean - variance (bcv) relationship, as well as colors and annotates genes of interest.
}
\examples{
plot_bcv(annotated_fData)
}
\keyword{plot_bcv}
