% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CoGAPS_ChiSq.R
\name{plot_CoGAPS_ChiSq}
\alias{plot_CoGAPS_ChiSq}
\title{Plots the mean ChiSq values of a set of CoGAPS results}
\usage{
plot_CoGAPS_ChiSq(CoGAPS_set)
}
\arguments{
\item{CoGAPS_res_set}{a CoGAPS result set.}
}
\value{
A ggplot2 object
}
\description{
plot_CoGAPS_ChiSq() plots the mean ChiSq values of a set of CoGAPS results.
}
\examples{
plot_CoGAPS_ChiSq(myCoGAPSres)
}
\keyword{plot_CoGAPS_ChiSq}
