% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_dendrogram.R
\name{plot_matrix_dendrogram}
\alias{plot_matrix_dendrogram}
\title{Plot a clustered dendrogram}
\usage{
plot_matrix_dendrogram(matrix, scale = "row", cluster = "col",
  clustering_distance = "euclidean", clustering_method = "complete")
}
\arguments{
\item{matrix}{a matrix of values}

\item{scale}{"rows" or "cols" or "none"; indicates whether values should be centered and scaled in the 
row direction or the column direction or none.}

\item{cluster}{"rows" or "cols"; whether to cluster rows or columns.}

\item{clustering_distance}{distance measure to be used. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method}{clustering method to be used. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}
}
\value{
a ggplot2 object
}
\description{
Draws a clustered heatmap using clustering methods implemented in the popular R package pheatmap. 
Returns a ggplot2 object
}
\examples{
plot_matrix_heatmap(matrix)
}
\keyword{plot_matrix_heatmap}
