% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_heatmap.R
\name{plot_matrix_heatmap}
\alias{plot_matrix_heatmap}
\title{Plot a clustered heatmap}
\usage{
plot_matrix_heatmap(matrix, scale = "row", limits = NULL,
  color_scale_type = "diverge", color_scale = c("blue", "lightyellow",
  "red"), scale_midpoint = 0, cluster_rows = T, cluster_cols = T,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  clustering_distance_cols = "euclidean",
  clustering_method_cols = "complete", square_tiles = F)
}
\arguments{
\item{matrix}{a matrix of values}

\item{scale}{"rows" or "cols" or "none"; indicates whether values should be centered and scaled in the 
row direction or the column direction or none.}

\item{limits}{minimum and maximum values to define the range of the color scale}

\item{color_scale_type}{"diverge", "seq" or "rescale"}

\item{color_scale}{vector of colors to use for color scale; different options for color_scale_type requires
different number of colors - "diverge" (3 colors), "seq" (2 colors), "rescale" (3 colors)}

\item{scale_midpoint}{numeric; value of midpoint for "diverge" and "rescale" color scales}

\item{cluster_rows}{logical; if rows should be clustered}

\item{cluster_cols}{logical; if columns should be clustered}

\item{clustering_distance_rows}{distance measure used in clustering rows. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method_rows}{clustering method used for rows. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{clustering_distance_cols}{distance measure used in clustering columns. Possible values are the same as
clustering_distance_rows.}

\item{clustering_method_cols}{clustering method used for columns. Accepts the same values as 
clustering_method_rows.}

\item{square_tiles}{logical; if tiles should be made squrare}
}
\value{
a ggplot2 object
}
\description{
Draws a clustered heatmap using clustering methods implemented in the popular R package pheatmap. Returns
a ggplot2 object
}
\examples{
plot_matrix_heatmap(matrix)
}
\keyword{plot_matrix_heatmap}
