% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_exprs_heatmap.R
\name{plot_mean_exprs_heatmap}
\alias{plot_mean_exprs_heatmap}
\title{Plot mean expression heatmap}
\usage{
plot_mean_exprs_heatmap(genes, cds, group, scale = NULL,
  cluster_groups = F, order_subgroups = F,
  cluster_groups_distance = "euclidean",
  cluster_groups_method = "complete", cluster_genes = F,
  cluster_genes_vector = NULL, cluster_genes_distance = "correlation",
  cluster_genes_method = "complete", color_scale = c("midnightblue",
  "gold", "grey50"), limits = NULL, square_tiles = T)
}
\arguments{
\item{genes}{a vector of gene name(s) to plot e.g. c("Actb", "Aldoa")}

\item{cds}{a monocle CellDataSet object}

\item{group}{column name in pData(cds) to group the bar plots by on the horizontal axis.}

\item{scale}{scaling to be done for each gene; possible values are "z-score" and "rescale"; "z-score"
scales the values to the z-scores; "rescale" scales the values to have a minimum of 0 and maximum of 1.}

\item{cluster_groups}{logical; if groups should be clustered.}

\item{order_subgroups}{logical; if subgroups should be ordered according to clustering of first subgroup. 
Subgroups can be specified in a string with separation by " - ".}

\item{cluster_groups_distance}{distance measure used in clustering groups. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{cluster_groups_method}{clustering method used for groups. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{cluster_genes}{logical; if genes should be clustered.}

\item{cluster_genes_vector}{a vector specifying the order of genes. If NULL, genes will be clustered
according to distance and method specified below.}

\item{cluster_genes_distance}{distance measure used in clustering genes. Possible values are the
same as cluster_groups_distance().}

\item{cluster_genes_method}{clustering method used for genes. Possible values are the
same as cluster_groups_method().}

\item{color_scale}{color scale vector for heatmap of length 3 for low, high, and na values respectively;
if null, limits of color scale will be determined by the squish() function}

\item{limits}{limits of the color scale for the heatmap.}

\item{square_tiles}{logical; if tiles should be made square}
}
\value{
a ggplot object
}
\description{
plot_mean_exprs_heatmap() plots a heatmap showing the mean expression of a gene in a monocle
CellDataSet object.
}
\examples{
plot_mean_exprs_heatmap(c("Actb","Aldoa"), cds=dat)
}
\keyword{plot_pattern_cor_heatmap}
