% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordered_marker_heatmap.R
\name{plot_ordered_marker_heatmap}
\alias{plot_ordered_marker_heatmap}
\title{Plot heatmap of marker genes ordered by cell group}
\usage{
plot_ordered_marker_heatmap(cds, marker_genes, group_vector, scale = T,
  cluster_rows = T, cluster_cols = T,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  clustering_distance_cols = "euclidean",
  clustering_method_cols = "complete", column_annotation,
  limits = c(-2, 2), show_rownames = T, show_colnames = F)
}
\arguments{
\item{cds}{CellDataSet object}

\item{marker_genes}{dataframe containing 2 columns: (1) gene symbol (2) cell group e.g. CellType}

\item{group_vector}{a vector classifying cells e.g. genotype or condition}

\item{scale}{"rows" or "cols" or "none"; indicates whether values should be centered and scaled in the 
row direction or the column direction or none.}

\item{cluster_rows}{logical; if rows should be clustered}

\item{cluster_cols}{logical; if columns should be clustered}

\item{clustering_distance_rows}{distance measure used in clustering rows. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method_rows}{clustering method used for rows. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{clustering_distance_cols}{distance measure used in clustering columns. Possible values are the same as
clustering_distance_rows.}

\item{clustering_method_cols}{clustering method used for columns. Accepts the same values as 
clustering_method_rows.}

\item{column_annotation}{column names in pData(cds) for annotating columns of heatmap}

\item{limits}{limits of color scale of heatmap}

\item{show_rownames}{logical; if rownames of heatmap should be displayed}

\item{show_colnames}{logical; if colnames of heatmap should be displayed}
}
\value{
A pheatmap object
}
\description{
plot_ordered_marker_heatmap() plots a heatmap with genes ordered by cell group.
}
\examples{
plot_ordered_marker_heatmap(dat, marker_genes = myMarkers, group_vector = "CellType")
}
\keyword{plot_ordered_marker_heatmap}
