% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_heatmap.R
\name{plot_pathway_heatmap}
\alias{plot_pathway_heatmap}
\title{Plot pathway heatmap}
\usage{
plot_pathway_heatmap(path_table = NULL, path_res_df,
  color_scale = c("blue", "lightyellow", "red"), cluster_row = T,
  cluster_col = T, mark_sig = T, sig_para = "pval",
  sigCutoff = 0.05, clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  clustering_distance_cols = "euclidean",
  clustering_method_cols = "complete", flip_axes = F,
  square_tiles = T)
}
\arguments{
\item{path_table}{a dataframe of values}

\item{path_res_df}{dataframe should contain columns "CellType", "pathway", "NES";can be output from 
the fgsea() function}

\item{color_scale}{vector of colors to use for color scale; different options for color_scale_type requires
different number of colors - "diverge" (3 colors), "seq" (2 colors), "rescale" (3 colors)}

\item{mark_sig}{logical; whether to mark statistically significant results with a "*".}

\item{sig_para}{parameter for statistical significance.}

\item{sigCutoff}{cutoff for statistical significance.}

\item{clustering_distance_rows}{distance measure used in clustering rows. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method_rows}{clustering method used for rows. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{clustering_distance_cols}{distance measure used in clustering columns. Possible values are the same as
clustering_distance_rows.}

\item{clustering_method_cols}{clustering method used for columns. Accepts the same values as 
clustering_method_rows.}

\item{flip_axes}{logical; if x- and y-axes should be flipped}

\item{square_tiles}{logical; if tiles should made into squares}

\item{cluster_rows}{logical; if rows should be clustered}

\item{cluster_cols}{logical; if columns should be clustered}
}
\value{
A ggplot2 object
}
\description{
plot_pathway_heatmap() plots the results of output from the fgsea() function.
}
\examples{
plot_pathway_heatmap(path_table=NULL, path_res_df=my_pathway_enrichment)
}
\keyword{plot_pathway_heatmap()}
