% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pattern_cor_dendro.R
\name{plot_pattern_cor_dendrogram}
\alias{plot_pattern_cor_dendrogram}
\title{Plot pattern correlation dendrogram}
\usage{
plot_pattern_cor_dendrogram(CoGAPS_res_set = NULL, Pattern_set,
  pattern_subset = NULL, annotation, group_variable, group_vector,
  order_subgroups = F, cluster = "patterns",
  clustering_distance = "correlation", clustering_method = "complete")
}
\arguments{
\item{CoGAPS_res_set}{a CoGAPS result set. If set to NULL, a Pmeans matrix may be supplied directly as input
for the Pattern_set parameter.}

\item{Pattern_set}{a string specifying the name of a pattern set (e.g. "nP20") in a CoGAPS_res_set. 
Alternatively, if CoGAPS_res_set is set to NULL, a Pmeans matrix may be supplied directly.}

\item{pattern_subset}{a vector specifying a subset of the pattern set to be plotted. Defaults to NULL.}

\item{annotation}{an annotation dataframe. Must contain a group_variable column with values corresponding to 
those in group_vector.}

\item{group_variable}{column name of column in annotation object with values corresponding to those 
in group_vector.}

\item{group_vector}{a vector of with pre-specified values corresponding to those in the group_variable in
the annotation object.}

\item{order_subgroups}{logical; if subgroups should be ordered according to clustering of first subgroup. 
Subgroups can be specified in a string with separation by " - ". Only valid if cluster is set to "groups".}

\item{cluster}{what to cluster. Possible values are "patterns" or "groups".}

\item{clustering_distance}{distance measure used in clustering. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{clustering_method}{clustering method used. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}
}
\value{
a ggplot object
}
\description{
plot_pattern_cor_dendro() plots a dendrogram showing the clustering of the correlation of a CoGAPS pattern 
set with a grouping variable
}
\examples{
plot_pattern_cor_heatmap(CoGAPS_res_set= my_CoGAPS_res, Pattern_set="nP30",  annotation=pData(cds), 
group_variable="genotype", group_vector=genotype_vector, order_subgroups=T, cluster= "groups",
clustering_distance = "correlation", clustering_method = "complete")
}
\keyword{plot_pattern_cor_dendro}
