% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pattern_cor_heatmap.R
\name{plot_pattern_cor_heatmap}
\alias{plot_pattern_cor_heatmap}
\title{Plot pattern correlation heatmap}
\usage{
plot_pattern_cor_heatmap(CoGAPS_res_set = NULL, Pattern_set,
  pattern_subset = NULL, annotation, compute_groups_by, group_vector,
  cluster_groups = F, order_subgroups = F,
  cluster_groups_distance = "correlation",
  cluster_groups_method = "complete", cluster_patterns = F,
  cluster_patterns_vector = NULL,
  cluster_patterns_distance = "correlation",
  cluster_patterns_method = "complete", limits = NULL,
  square_tiles = T)
}
\arguments{
\item{CoGAPS_res_set}{a CoGAPS result set. If set to NULL, a Pmeans matrix may be supplied directly as input
for the Pattern_set parameter.}

\item{Pattern_set}{a string specifying the name of a pattern set (e.g. "nP20") in a CoGAPS_res_set. 
Alternatively, if CoGAPS_res_set is set to NULL, a Pmeans matrix may be supplied directly.}

\item{pattern_subset}{a vector specifying a subset of the pattern set to be plotted. Defaults to NULL.}

\item{annotation}{an annotation dataframe. If group_column is set to NULL, must contain logical columns with 
column names corresponding to values in group_vector. Each logical column (1 or 0) should indicate whether an
annotation row is of the value (column name).}

\item{group_vector}{a vector of with pre-specified values corresponding to those in group_column in
the annotation object.}

\item{cluster_groups}{logical; if groups should be clustered.}

\item{order_subgroups}{logical; if subgroups should be ordered according to clustering of first subgroup. 
Subgroups can be specified in a string with separation by " - ".}

\item{cluster_groups_distance}{distance measure used in clustering groups. Possible values are "correlation",
"euclidean", "maximum", "manhattan", "canberra", "binary", or "minkowski" (defined in dist() function.)}

\item{cluster_groups_method}{clustering method used for groups. Accepts the same values as hclust: "ward.D", 
"ward.D2", "single", "complete", "average" (=UPGMA), "mcquitty" (=WPGMA), "median" (=WPGMC) or 
"centroid" (=UPGMC).}

\item{cluster_patterns}{logical; if patterns should be clustered.}

\item{cluster_patterns_vector}{a vector specifying the order of patterns. If NULL, patterns will be clustered
according to distance and method specified below.}

\item{cluster_patterns_distance}{distance measure used in clustering patterns. Possible values are the
same as cluster_groups_distance().}

\item{cluster_patterns_method}{clustering method used for patterns. Possible values are the
same as cluster_groups_method().}

\item{limits}{limits of the color scale for the heatmap.}

\item{square_tiles}{logical; if tiles should be made square}

\item{group_column}{column name of column in annotation object with values corresponding to those 
in group_vector to use for computing logical columns. Defaults to NULL}
}
\value{
a ggplot object
}
\description{
plot_pattern_cor_heatmap() plots a heatmap showing the correlation of a CoGAPS pattern set with a grouping
variable
}
\examples{
plot_pattern_cor_heatmap(CoGAPS_res_set = my_CoGAPS_res, Pattern_set = "nP30",annotation=pData(cds), 
group_column=NULL, group_vector=genotype_vector, cluster_groups=T, order_subgroups=T, 
cluster_groups_distance = "correlation", cluster_groups_method = "complete", cluster_patterns= T, 
cluster_patterns_distance = "correlation", cluster_patterns_method = "complete")
}
\keyword{plot_pattern_cor_heatmap}
