% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patternSet_dim.R
\name{plot_patternSet_dim}
\alias{plot_patternSet_dim}
\title{Plot a set of patterns on a dimensionality reduction plot}
\usage{
plot_patternSet_dim(CoGAPS_res_set, Pattern_set, annotation, dim_reduction,
  facet_wrap_by = NULL)
}
\arguments{
\item{CoGAPS_res_set}{a CoGAPS result set. If set to NULL, a Pmeans matrix may be supplied directly as input
for the Pattern_set parameter.}

\item{Pattern_set}{a string specifying the name of a pattern set (e.g. "nP20") in a CoGAPS_res_set. 
Alternatively, if CoGAPS_res_set is set to NULL, a Pmeans matrix may be supplied directly.}

\item{annotation}{an annotation dataframe. Must contain columns with column names corresponding to values
in the group_vector.}

\item{dim_reduction}{a dataframe specifying the coordinates of a dimensionality reduction method (e.g. 
t-SNE, UMAP etc.). Each column should specify one dimension. Rownames should correspond to the rownames
of the annotation object. Colnames will be used as axis labels.}

\item{facet_wrap_by}{a string specifying variable for facet_wrap; defaults to NULL.}
}
\value{
a list of ggplot objects
}
\description{
plot_patternSet_dim() plots Pmeans of a CoGAPS pattern set on a dimensonality reduction plots.
}
\examples{
plot_patternSet_dim(CoGAPS_res_set= my_CoGAPS_res, Pattern_set="nP30",  annotation=pData(cds), 
dim_reduction = UMAP_proj)
}
\keyword{plot_pattern_dim}
