% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pheatmap.R
\name{plot_pheatmap}
\alias{plot_pheatmap}
\title{Plot pheatmap object}
\usage{
plot_pheatmap(pheatmap_object)
}
\arguments{
\item{pheatmap_object}{object containing output from the pheatmap function. Created by assigning output of 
pheatmap function to an object (i.e. pheatmap_object <- pheatmap(matrix))}
}
\value{
A heatmap created by pheatmap
}
\description{
Function for plotting a heatmap from an object created by assignment of output from the pheatmap
function in the popular R package pheatmap
}
\examples{
pheatmap_object <- pheatmap(matrix)
plot_pheatmap(pheatmap_object)
}
\keyword{plot_pheatmap}
