% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sparse_pca_var.R
\name{plot_sparse_pca_var}
\alias{plot_sparse_pca_var}
\title{Plot variance explained by sparse PCA}
\usage{
plot_sparse_pca_var(sparse_pca_obj, n_pcs)
}
\arguments{
\item{sparse_pca_obj}{a sparse pca object created by the cellRangerRkit sparse_pca() function}

\item{n_pcs}{number of PCs to plot; defaults to NULL; if null, all the PCs will be plotted}
}
\value{
A ggplot object
}
\description{
plot_sparse_pca_var() plots the variance explained by sparse PCA (implemented by cellrangerRkit) as a
line graph.
}
\examples{
plot_sparse_pca_var(sparse_pca_obj, n_pcs = 40)
}
\keyword{plot_sparse_pca_var}
