% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_SD_SE.R
\name{rowSD}
\alias{rowSD}
\title{Get standard deviation or standard error for rows in a matrix}
\usage{
rowSD(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
a numerical vector
}
\description{
rowSD() and rowSE() calculates row-wise standard deviation and standard error of the mean for a matrix
}
\examples{
row_Std_Dev <- rowSD(matrix)
}
\keyword{rowSD(),}
\keyword{rowSE()}
