% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_SE.R
\name{rowSE}
\alias{rowSE}
\title{Get standard error of the mean for rows in a matrix}
\usage{
rowSE(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
a numerical vector
}
\description{
rowSE() calculates row-wise standard error of the mean for a matrix
}
\examples{
row_Std_Err <- rowSE(matrix)
}
\keyword{rowSE()}
