% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_pca.R
\name{sparse_pca}
\alias{sparse_pca}
\title{Maintain sparsity and run a fast PCA algorithm}
\usage{
sparse_pca(matrix, n_pcs, mu = NULL, s = NULL, center_scale = TRUE)
}
\arguments{
\item{matrix}{a sparse matrix}

\item{n_pcs}{number of principal components to compute}

\item{mu}{column means}

\item{s}{column standard deviations}

\item{center_scale}{logical; whether to perform centering and scaling}
}
\value{
an irlba pca object
}
\description{
sparse_pca() runs a fast PCA algorithm on a sparse matrix. Adapted from CellRangerRkit.
}
\examples{
pca_object <- sparse_pca(exprs(dat), n_pcs = 40)
}
\keyword{sparse_pca()}
