% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_aggr_mtx.R
\name{split_aggr_mtx}
\alias{split_aggr_mtx}
\title{Split a cellranger-aggregated matrix into multiple matrices based on barcodes}
\usage{
split_aggr_mtx(aggr_mtx)
}
\arguments{
\item{aggr_mtx}{an expression matrix derived from the cellranger aggr function}
}
\value{
a dgTMatrix
}
\description{
split_aggr_matrix() splits an expression matrix derived from a directory of 10X Genomics scRNAseq 
data aggregated from multiple samples created using the cellranger pipeline. Based on the unique numeric 
value assigned to barcodes of each sample by the cellranger aggr function.
}
\examples{
sample_raw_mtx <- split_aggr_mtx(raw_mtx)
}
\keyword{split_aggr_matrix}
