% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_cds.R
\name{subset_cds}
\alias{subset_cds}
\title{Subset cds by cell group}
\usage{
subset_cds(cell_group, cds, group, min_expr = NULL, read_type = "UMIs")
}
\arguments{
\item{cell_group}{string specifying specific cell group e.g. "Rods"}

\item{cds}{a CellDataSet object used in the monocle package}

\item{group}{column name in pData(cds) specifying cell group that each cell belong to e.g. "CellType"}

\item{min_expr}{expression threshold to use for detectGenes() function}

\item{read_type}{string specifying type of reads in expression matrix; defaults to "UMIs"}
}
\value{
CellDataSet object
}
\description{
subset_cds() subsets a CellDataSet object for a particular cell group and prepares the new CellDataSet
subset for the monocle differentialGeneTest() function.
}
\examples{
subset_cds("Rods", dat, group = "CellType", min_expr = 1)
}
\keyword{gene_barplot()}
