% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDrops.R
\name{testDrops}
\alias{testDrops}
\title{Determine whether droplets are cells or blank drops in an expression matrix}
\usage{
testDrops(raw_exprs_mtx, lower_UMI_threshold = 100,
  upper_UMI_threshold = 500, test.ambient = T, FDR = 0.01,
  niters = 10000)
}
\arguments{
\item{raw_exprs_mtx}{a raw expression matrix created with the cellranger pipeline; rows are genes and 
columns are barcodes}

\item{lower_UMI_threshold}{numeric; parameter for the testDrops function. All barcodes with total UMI counts
below this value are considered empty droplets (i.e. not cells)}

\item{upper_UMI_threshold}{numeric; parameter for the testDrops function. All barcodes with total UMI counts
above this value are considered cells}

\item{test.ambient}{A logical scalar indicating whether results should be returned for barcodes with totals 
less than or equal to lower_UMI_threshold.}

\item{FDR}{numeric; parameter for the testDrops function. FDR value to use in the testDrops function when
statistically determining cells vs non-cells}

\item{niters}{An integer scalar specifying the number of iterations to use for the Monte Carlo p-value 
calculations.}
}
\value{
a testDrops object with a results dataframe, cell_barcodes, and blank_barcodes
}
\description{
testDrops() is a wrapper function around the emptyDrops() function in the excellent SoupX package. Returns
a list object consisting of a results dataframe from the emptyDrops() function, as well as barcodes 
of cells and empty droplets.
}
\examples{
testDrops_res <- testDrops(raw_exprs_mtx, lower_UMI_threshold=100, upper_UMI_threshold=500, 
test.ambient=T, FDR=0.01, niters=10000)
}
\keyword{testDrops}
