% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bcv.R
\name{annotate_bcv}
\alias{annotate_bcv}
\title{Calculate and annotate biological coefficent of variation}
\usage{
annotate_bcv(cds)
}
\arguments{
\item{cds}{a monocle CellDataSet object}
}
\value{
an annotated fData dataframe
}
\description{
annotate_bcv() calculates the biological coefficent of variation from a give cds and returns an annotated
fData dataframe.
}
\examples{
annotate_bcv(dat)
}
\keyword{annotate_bcv}
