% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mt_exprs.R
\name{calculate_mt_exprs}
\alias{calculate_mt_exprs}
\title{Calculate mitochondrial gene expression}
\usage{
calculate_mt_exprs(cds, genome = "mouse", return_prop = T)
}
\arguments{
\item{cds}{monocle CellDataSet object}

\item{genome}{genome to which transcripts are aligned to. Accepts either "mouse" or "human".}

\item{return_prop}{logical; whether to return the value of mitochondrial transcripts as 
a proportion of total transcripts for each cell; resulting output will be the input CDS with
the values of mt_RNAs and mt_prop attached to the corresponding pData. If FALSE, will only
return the absolute number of transcripts per cell as a vector. Defaults to TRUE.}
}
\value{
A vector or a cds (see above)
}
\description{
calculate_mt_exprs() calculates the expression of mitochondrial genes for a monocle CellDataSet
object.
}
\examples{
calculate_mt_exprs
}
\keyword{mt_RNAs}
\keyword{mt_genes,}
