% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cds_vglm.R
\name{fit_cds_vglm}
\alias{fit_cds_vglm}
\title{Fit vector generalized linear model to a CellDataSet object}
\usage{
fit_cds_vglm(cds, modelFormulaStr, test_genes = c("Actb"))
}
\arguments{
\item{cds}{a monocle CellDataSet object}

\item{modelFormulaStr}{model formula string specifying the model to fit for the genes}

\item{test_genes}{genes to test the model on; test_genes must be expressed in cds object.}
}
\description{
fit_cds_vglm() is a wrapper function around the monocle::fitModel() function that fits a vglm for each 
gene in a CellDataSet object and return a dataframe with the estimate, standard error, z value and
p-value for each coefficient specified in the model formula string.
}
\examples{
find_cds_vglm(dat)
}
\keyword{fit_cds_vglm()}
