% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_barplot.R
\name{gene_barplot}
\alias{gene_barplot}
\title{Plot mean gene expression as bar plot}
\usage{
gene_barplot(gene_to_plot, cds, group = "genotype", color = "genotype",
  color_scale = c("Red", "Black"), plot_trend = F)
}
\arguments{
\item{gene_to_plot}{official gene symbol of a gene to plot e.g. "Actb"}

\item{cds}{a CellDataSet object e.g. used in the monocle package}

\item{group}{column name in pData(cds) to group the bar plots by on the horizontal axis.}

\item{color}{column name in pData(cds) to color the bar plots with.}

\item{color_scale}{a vector of colors to color barplots and error bars with.}
}
\value{
A ggplot2 object
}
\description{
gene_barplot() plots the mean expression of a given gene as a bar plot with standard error bars.
}
\examples{
gene_barplot("Actb", dat)
}
\keyword{gene_barplot()}
