% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_jitterplot.R
\name{gene_jitterplot}
\alias{gene_jitterplot}
\title{Plot expression of cells as jitter plot}
\usage{
gene_jitterplot(genes, cds, group = "genotype", color = NULL,
  cell_size = 0.75, min_expr = NULL, nrow = NULL, ncol = 1,
  panel_order = NULL, plot_trend = FALSE, color_trend = "orange",
  label_by_short_name = TRUE, relative_expr = TRUE)
}
\arguments{
\item{genes}{a vector of gene name(s) to plot e.g. c("Actb", "Aldoa")}

\item{cds}{a CellDataSet object e.g. used in the monocle package}

\item{group}{column name in pData(cds) to group cells by on the horizontal axis.}

\item{color}{column name in pData(cds) to color cells with.}

\item{cell_size}{the size (in points) of each cell used in the plot}

\item{min_expr}{the minimum (untransformed) expression level to use in plotted the genes.}

\item{nrow}{the number of rows used when laying out the panels for each gene's expression}

\item{ncol}{the number of columns used when laying out the panels for each gene's expression}

\item{plot_trend}{logical;whether to plot a trend line across groups in plot(s).}

\item{color_trend}{color to use for trend line.}

\item{label_by_short_name}{logical;label figure panels by gene_short_name (TRUE) or feature id (FALSE)}

\item{relative_expr}{logical;Whether to transform expression into relative values}
}
\value{
A ggplot2 object
}
\description{
gene_jitterplot() plots the expression of a given gene as a jitter plot for a CellDataSet object.
}
\examples{
gene_jitterplot(c("Actb", "Aldoa"), cds)
}
\keyword{gene_jitterplot}
