% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drops_diagnostics.R
\name{plot_drops_diagnostics}
\alias{plot_drops_diagnostics}
\title{Plot diagnostic plot for testDrops results}
\usage{
plot_drops_diagnostics(testDrops_res)
}
\arguments{
\item{testDrops_res}{a testDrops object create by a testDrops function}
}
\value{
a ggplot2 object
}
\description{
Plots a diagnostic plot for a testDrops result. The x-axis represents the total number of UMIs per cell,
while the y-axis represents the negative log probability that a droplet is a cell.
}
\examples{
plot_drops_diagnostics(testDrops_res)
}
\keyword{plot_drops_diagnostics}
