% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_SD.R
\name{rowSD}
\alias{rowSD}
\title{Get standard deviation for rows in a matrix}
\usage{
rowSD(matrix)
}
\arguments{
\item{matrix}{a matrix}
}
\value{
a numerical vector
}
\description{
rowSD() calculates row-wise standard deviation for a matrix
}
\examples{
row_Std_Dev <- rowSD(matrix)
}
\keyword{rowSD()}
