% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_pca.R
\name{sparse_pca}
\alias{sparse_pca}
\title{Perform sparse pca on a matrix}
\usage{
sparse_pca(x, n_pcs, mu = NULL, s = NULL, center_scale = TRUE)
}
\arguments{
\item{x}{a matrix of values to perform dimensionality reduction on; by default, rows are genes
and columns are cells}

\item{n_pcs}{number of prinicpal components to compute}

\item{mu}{column means}

\item{s}{column standard deviations}

\item{center_scale}{perform centering and scaling}
}
\value{
A list containing
"x" - The rotated data matrix where rows are barcodes and columns are PCs
"sdev" - the standard deviations of the principal components (i.e., sqrt of eigvals of the covariance matrix)
"rotation" - The loadings (eigenvectors) where each column is a PC
"tot_var" - The total variation in the scaled and centered matrix (this is also the effective rank of the matrix)
"var_pcs" - The proportion of variance explained by each principle comoponent
}
\description{
sparse_pca() performs a fast pca on a matrix while maintaining sparsity.
}
\examples{
sparse_pca(mtx, n_pcs=10)
}
\keyword{sparse_pca}
