% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_tsne.R
\name{spectral_tsne}
\alias{spectral_tsne}
\title{Reduce dimensions of a matrix with pca and tSNE}
\usage{
spectral_tsne(matrix, log_matrix = TRUE, prcomp_object = NULL,
  dims = 1:10, perplexity = 30)
}
\arguments{
\item{matrix}{a matrix of values to perform dimensionality reduction on}

\item{log_matrix}{if log10 transformation is to be performed on the matrix; defaults to TRUE}

\item{dims}{dimensions from the prinicpal component analysis to use; defaults to 1:10 (i.e. 1st to 10th
principal components)}

\item{perplexity}{numeric; perplexity parameter for tSNE; defaults to 30}

\item{prcomp_obj}{a principal component analysis object produced by the prcomp or prcomp_irlba functions;
if no object is supplied, irlba_prcomp will be run on the matrix to return 50 dimensions; defaults to NULL}
}
\value{
A matrix with two columns containing coordinates of each row for two dimensions respectively
}
\description{
spectral_tsne() performs dimensionality reduction using the package Rtsne's wrapper for the 
C++ implementation of Barnes-Hut t-Distributed Stochastic Neighbor Embedding. If no prcomp_object 
is supplied, principal component analysis will be performed using the irlba package.
}
\examples{
spectral_tsne(matrix, log_matrix=TRUE, prcomp_object=NULL, dims=1:10, perplexity=30)
}
\keyword{spectral_tsne}
