% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_umap.R
\name{spectral_umap}
\alias{spectral_umap}
\title{Reduce dimensions of a matrix with pca and UMAP}
\usage{
spectral_umap(matrix, log_matrix = TRUE, prcomp_object = NULL,
  dims = 1:10)
}
\arguments{
\item{matrix}{a matrix of values to perform dimensionality reduction on}

\item{log_matrix}{if log10 transformation is to be performed on the matrix; defaults to TRUE}

\item{dims}{dimensions from the prinicpal component analysis to use; defaults to 1:10 (i.e. 1st to 10th
principal components)}

\item{prcomp_obj}{a principal component analysis object produced by the prcomp or prcomp_irlba functions;
if no object is supplied, irlba_prcomp will be run on the matrix to return 50 dimensions; defaults to NULL}
}
\value{
A matrix with two columns containing coordinates of each row for two dimensions respectively
}
\description{
spectral_umap() performs dimensionality reduction using the package monocle wrapper function around
the python implementation of UMAP (Uniform Manifold Approximation and Projection). If no prcomp_object 
is supplied, principal component analysis will be performed using the irlba package.
}
\examples{
spectral_umap(matrix, log_matrix=TRUE, prcomp_object=NULL, dims=1:10)
}
\keyword{spectral_umap}
