% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cds_exprs.R
\name{summarize_cds_exprs}
\alias{summarize_cds_exprs}
\title{Get gene-level summary statistics for CellDataSet object}
\usage{
summarize_cds_exprs(cds, group, gene_short_name = T)
}
\arguments{
\item{cds}{a CellDataSet object}

\item{group}{one or more column(s) specifying groups to obtain summary statistics for e.g. CellType}

\item{gene_short_name}{logical; whether to attach gene_short_name as a column to output dataframe}
}
\value{
a dataframe
}
\description{
summarize_cds_exprs() takes a monocle CellDataSet object as input and obtain summary statistics 
for each gene for specified groups in the dataset.
}
\examples{
dat_summary <- summarize_cds_exprs(dat, c("CellType"))
}
\keyword{summarize_cds_exprs()}
